move dev control types to shared directory
